OpenSSL: fix warnings related to sign

+++ openvpn/openssl/ssl/sslctx.hpp
@@ -736,7 +736,7 @@ namespace openvpn {
 	    throw ssl_external_pki("OpenSSL: could not obtain signature");
 
 	  /* decode base64 signature to binary */
-	  const int len = RSA_size(rsa);
+	  const size_t len = RSA_size(rsa);
 	  Buffer sig(to, len, false);
 	  base64->decode(sig, sig_b64);
 
@@ -1089,7 +1089,7 @@ namespace openvpn {
 	      ASN1_STRING *val = X509_NAME_ENTRY_get_data(ent);
 	      unsigned char *buf;
 	      buf = (unsigned char *)1; // bug in OpenSSL 0.9.6b ASN1_STRING_to_UTF8 requires this workaround
-	      const int len = ASN1_STRING_to_UTF8(&buf, val);
+	      const size_t len = ASN1_STRING_to_UTF8(&buf, val);
 	      if (len > 0)
 		{
 		  if (std::strlen((char *)buf) == len)
@@ -1114,7 +1114,7 @@ namespace openvpn {
 			  if (BIO_write(bio, &nullc, 1) == 1)
 			    {
 			      char *str;
-			      const long len = BIO_get_mem_data(bio, &str);
+			      const size_t len = BIO_get_mem_data(bio, &str);
 			      if (std::strlen(str) == len)
 				ret = str;
 			    }
